split py35 tests
